requirements.txt
python-redmine==0.6.2